The examples have been updated
