--- examples/createGroupAndSendMessage.py
@@ -1,30 +1,0 @@
-from os import environ
-
-from whatsapp_api_client_python import GreenAPI
-
-# First you need to set the environment variables
-ID_INSTANCE = environ['ID_INSTANCE']
-API_TOKEN_INSTANCE = environ['API_TOKEN_INSTANCE']
-
-greenAPI = GreenAPI(ID_INSTANCE, API_TOKEN_INSTANCE)
-
-def main():
-    chatIds = [
-        "79001234567@c.us"
-    ]
-    resultCreate = greenAPI.groups.createGroup('GroupName', 
-        chatIds)
-
-    if resultCreate.code == 200:
-        print(resultCreate.data)
-        resultSend = greenAPI.sending.sendMessage(resultCreate.data['chatId'], 
-            'Message text')
-        if resultSend.code == 200:
-            print(resultSend.data)
-        else:
-            print(resultSend.error)    
-    else:
-        print(resultCreate.error)
-
-if __name__ == "__main__":
-    main()
